requirements.txt
ffmpeg-python
openai
